Pull Generic Request patch into its own file
This puts the http_generic_request into its own file.
